+++ aries-site/src/examples/components/layer/index.js
@@ -1,5 +1,6 @@
 export * from './LayerExample';
 export * from './LayerCenterExample';
-export * from './LayerFormExample';
 export * from './LayerSideDrawerExample';
-export * from './LayerStickyFormExample';
+export * from './LayerStickyExample';
+export * from './LayerStickyScrollExample';
+export * from './MonitorFormExample';
